pyproject.toml
[project]
name = "transactions-categorizer"
version = "0.1.0"
description = "Parse credit card statement PDFs and categorize transactions using local LLMs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic==2.10.6",
    "loguru==0.7.3",
    "httpx==0.28.1",
    "pdf2image==1.17.0",
    "pillow==11.1.0",
    "pytesseract==0.3.13",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "ruff==0.9.3",
]

[project.scripts]
categorize = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
package = true
